Extracting common file matchers behavior to FileMatcher class
